added useCopy functionality
