+++ package.json
@@ -73,6 +73,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-editor-js": "^2.1.0",
+    "react-hotkeys-hook": "^4.4.1",
     "react-intersection-observer": "^9.4.1",
     "react-mathlive": "^3.0.5-preview.1",
     "react-mentions": "^4.4.7",
