+++ src/components/editor/Editor.js
@@ -37,6 +37,7 @@ import useCriticalSections from '../../utils/useCriticalSections';
 import useLocking from '../../utils/useLocking';
 import useWebSocket from '../useWebSocket';
 import Modal from '../modal/Modal';
+import useCopy from '../../utils/useCopy';
 
 type Props = {
   status: 'inProgress' | 'inReview' | 'published',
@@ -70,6 +71,7 @@ function Editor({
   const { labelCriticalSections } = useCriticalSections({ blocks, enabled: get(currentPermissions, permissions.criticalSections, false) });
   const { checkLocks } = useLocking({ blocks, editor: editor.current, enabled: get(currentPermissions, permissions.locking, false) });
   useWebSocket({ articleId: get(article, 'id'), enabled: get(currentPermissions, permissions.webSockets, false) });
+  const popover = useCopy({ enabled: readOnly || false });
 
   // useEffect(() => {
   //   const defaultEditor = document.getElementById('editorjs');
@@ -78,7 +80,11 @@ function Editor({
   //   console.log('ref', defaultEditor);
   //   if (defaultEditor && readOnlyEditor) {
   //     readOnlyEditor.innerHTML = defaultEditor.innerHTML;
+  //     console.log('readonly', readOnlyEditor);
   //   }
+
+  //   const blockContents = document.getElementsByClassName('ce-block__content');
+  //   console.log('blockContents', blockContents);
   // }, []);
 
   useEffect(() => {
@@ -259,9 +265,10 @@ function Editor({
           position: 'relative',
           // display: 'none',
         }}
-      />
+      >
+        {popover}
+      </div>
       {/* <div
-        className='readonly-editor-wrapper'
         id="readonly-editorjs"
       /> */}
     </>
